% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcosw.R
\name{dcosw}
\alias{dcosw}
\title{The probability density function of the CosWeibull probability distribution.}
\usage{
dcosw(x, alpha, lambda)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{alpha}{Alpha parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the CosWeibull distribution.
}
\description{
The probability density function of the CosWeibull probability distribution.
}
\examples{
dcosw(1,1,1)
dcosw(1,1.5,2)
}

