% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcosw.R
\name{hcosw}
\alias{hcosw}
\title{The hazard function of the CosWeibull probability distribution.}
\usage{
hcosw(x, alpha, lambda)
}
\arguments{
\item{x}{Vector of quantiles..}

\item{alpha}{Alpha parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the CosWeibull distribution.
}
\description{
The hazard function of the CosWeibull probability distribution.
}
\examples{
hcosw(1,1.5,2)
hcosw(1,2,0.5)
}

