% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcosw.R
\name{qcosw}
\alias{qcosw}
\title{The quantile function of the CosWeibull probability distribution.}
\usage{
qcosw(p, alpha = 1, lambda, lower = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{alpha}{Alpha parameter.}

\item{lambda}{Lambda parameter.}

\item{lower}{Lower parameter.}

\item{log.p}{Log.p parameter.}
}
\value{
A vector with n observations of the CosWeibull distribution.
}
\description{
The quantile function of the CosWeibull probability distribution.
}
\examples{
qcosw(1, 1, 1, TRUE, FALSE)
qcosw(1, 1, 0.1, TRUE, FALSE)
}

