% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcosw.R
\name{rcosw}
\alias{rcosw}
\title{Generates random deviates from a CosWeibull probability distribution.}
\usage{
rcosw(n, alpha, lambda)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{alpha}{Alpha parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the CosWeibull distribution.
}
\description{
Generates random deviates from a CosWeibull probability distribution.
}
\examples{
rcosw(1000,1,1)
rcosw(1000,1,0.1)
}

