% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scosw.R
\name{scosw}
\alias{scosw}
\title{The survival function of the CosWeibull probability distribution.}
\usage{
scosw(x, alpha, lambda)
}
\arguments{
\item{x}{Vector of quantiles..}

\item{alpha}{Alpha parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the CosWeibull distribution.
}
\description{
The survival function of the CosWeibull probability distribution.
}
\examples{
hcosw(1,1,1)
hcosw(1,1.5,2)
}

